// src/app/entities/meal/model/meal.model.ts
import { MealIngredient } from './meal-ingredient.model';

export interface Meal {
  id: number;
  name: string;
  servings: number;
  ingredients: MealIngredient[];
  totalCalories: number;
  totalProtein: number;
  totalFat: number;
  totalCarbohydrates: number;
  totalPrice: number;
  caloriesPerServing: number;
  proteinPerServing: number;
  fatPerServing: number;
  carbohydratesPerServing: number;
  pricePerServing: number;
  totalSodiumMg?: number | null;
  totalPotassiumMg?: number | null;
  totalMagnesiumMg?: number | null;
  totalIronMg?: number | null;
  totalCalciumMg?: number | null;
  totalZincMg?: number | null;
  totalVitaminAMcg?: number | null;
  totalVitaminCMg?: number | null;
  totalVitaminDMcg?: number | null;
  totalVitaminEMg?: number | null;
  totalVitaminKMcg?: number | null;
  totalVitaminB1Mg?: number | null;
  totalVitaminB2Mg?: number | null;
  totalVitaminB6Mg?: number | null;
  totalVitaminB9Mcg?: number | null;
  totalVitaminB12Mcg?: number | null;
  sodiumMgPerServing?: number | null;
  potassiumMgPerServing?: number | null;
  magnesiumMgPerServing?: number | null;
  ironMgPerServing?: number | null;
  calciumMgPerServing?: number | null;
  zincMgPerServing?: number | null;
  vitaminAMcgPerServing?: number | null;
  vitaminCMgPerServing?: number | null;
  vitaminDMcgPerServing?: number | null;
  vitaminEMgPerServing?: number | null;
  vitaminKMcgPerServing?: number | null;
  vitaminB1MgPerServing?: number | null;
  vitaminB2MgPerServing?: number | null;
  vitaminB6MgPerServing?: number | null;
  vitaminB9McgPerServing?: number | null;
  vitaminB12McgPerServing?: number | null;
}